(cleanup) Revert changes to authConfig because they are untestable with current OAuth settings.
