+++ client/lib/socketConfig.js
@@ -15,7 +15,7 @@ export const socket = io.connect(connectionString, {
   reconnectionAttempts: 3,
 });
 
-export const socketActionMiddleware = (socket, store, next, action) => {
+export const socketActionMiddleware = (socket) => (store) => (next) => (action) => {
   let meta = action.meta;
   if (meta) {
     socket.emit(meta.event, { to: meta.to, entry: meta.entry });
